Strappy.CL

Documentation

data Comb

Combinator data type. A well formed combinator is either a primitive combinator (CNode) or the application of a combinator to another combinator (CApp). A not-yet-written combinator is represented by a CTerminal (not well-named). TODO: we should change the name of CTerminal to something more appropriate.

Constructors

CApp

Fields

lComb :: Comb

rComb :: Comb

cType :: Type

cAppDepth :: Int

cLabel :: Maybe String

the label of a combinator is a unique name for the subtree

CLeaf

Fields

cName :: String

cExpr :: Expr

cType :: Type

CInnerNode
a location in a tree

Fields

cType :: Type

CHole

Fields

cHole :: Hole

class ToComb a where

Methods

toComb :: a -> Comb

data Domain a

Constructors

Domain [a]

data Hole

Constructors

forall a . (Show a, Eq a, ToComb a) => Hole

Fields

holeType :: Type

holeDomain :: Domain a

cDepth :: Comb -> Int

Return the precalculated depth of a combinator.

mkAppDepth :: Comb -> Comb -> Int

Calcuate the depth of the combinator resulting from an application.

isCLeaf :: Comb -> Bool

type SynthComb = Either String Comb

app :: SynthComb -> SynthComb -> SynthComb

app' :: Comb -> Comb -> Comb

Returns the combinator resulting from the application of two combinators. Throws error if there is a type mismatch.

(<:>) :: SynthComb -> SynthComb -> SynthComb

Alias for app.

(<::>) :: SynthComb -> Comb -> SynthComb

show' :: Comb -> String

An alternative to show: if combinator is named and evaluates to a number or bool, show it an an evaluated expressions.

num2C :: Int -> Comb

Convert integer to a combinator.

dOp2C

Arguments

:: String
operator name

-> (Int -> Int -> Int)
operator

-> Comb

Convert a binary integer operator to an expression.

bool2C :: Bool -> Comb

Convert a boolean value to a combinator.

getType :: Comb -> Either String Type

Get the type of a combinator outside the type inference monad.

getAppType :: Comb -> Comb -> Either String Type

Returns the type of an application of two combinators. If there is a type mismatch, returns Left.

sub :: TyVar -> Type -> Comb -> Comb

Apply a substitution to a type inside a combinator.

getTvs :: Comb -> [TyVar]

Returns all type variables used in a combinator

comb2Expr :: Comb -> Expr

Returns expression associated with combinator.

filterCombinatorsByType :: [Comb] -> Type -> StateT Int [] Comb

reduceComb :: Comb -> Expr

Convert a combinator to an expression and reduce.

fitCombWithHoles :: Comb -> Comb

data Dir

Constructors

L

R

type TrPtr = Seq Dir

getHoles :: Comb -> [(Hole, TrPtr)]

getHoles takes a combinator and returns a list of holes and their locations (i.e. pointers of type TrPtr to their locations in a tree).

productOfHoles :: [Hole] -> [[Comb]]

resolveCombHoles :: Comb -> [Comb]

Given a combinator with holes, this function returns a lazy list of all combinators formed by extending each hole to one of its domain values.

intIntervalDomain :: Int -> Int -> Domain Int

intIntervalHole :: Int -> Int -> Hole

intHole :: Hole

boolHole :: Hole