Strappy.Compress

Description

This module contains functions for compressing a set of combinators.

Documentation

type Index = CombMap Int

Record all unique combinators that satisfy the following:

incr :: Index -> Comb -> Index

When incrementing a tree: a ) if it has count 0 set count to 1 and increment children b) if it has count 1 set count to 2 and decrement children c) if it has count > 2 set count (+1) | When decrementing a tree: a) if it has count 1 set count to 0 and decrement children b) if it has count 2 or greater, set count (-1) and increment children

compress :: [Comb] -> Index

incr2 :: CombMap [Type] -> Comb -> Type -> State Int (CombMap [Type])

compress2 :: [(Type, Comb)] -> CombMap [Type]

getUniqueTrees :: Comb -> Index

Alternative compression scheme 1) index and count all the unique subexpressions in the library 2) the complexity of a set of expressions is the number of unique subexpression.