Strappy.EnumBF

Documentation

data Turn

Constructors

L

R

type Path = [Turn]

data CombBase

An ADT corresponding to a ``solution base.'' See Pearl Heuristics book Chapter 2.

Constructors

CombBase

Fields

comb :: Comb

path :: Maybe Path

a path to the current node of interest

value :: Double

data CombBaseTuple

Constructors

CombBaseTuple

Fields

combBase :: CombBase

tpVar :: Int

data BFState

Constructors

BFState

Fields

open :: MaxQueue CombBaseTuple

closed :: MaxQueue CombBaseTuple

isClosed :: CombBase -> Bool

hackMinGrammarVal :: Double

enumBF :: Grammar -> Int -> Type -> [CombBase]

Breadth-first AO enumeration of combinators with highest scores under the grammar.

enumBF' :: Grammar -> Int -> BFState -> BFState

Removes and expands the maximal entry in the OPEN list. The resulting elements are added either to the open list or to the closed list. This process is repeated until there are at least i elements in closed list.

expandToApp :: Grammar -> CombBase -> StateT Int [] CombBase

expand :: Grammar -> CombBase -> StateT Int [] CombBase

doTypesUnify :: Type -> Type -> Bool

doesCombHaveToType :: Comb -> Type -> State Int Bool

existCombWithToType :: [Comb] -> Type -> Bool