Strappy.Grammar

Documentation

data Grammar

Constructors

Grammar

Fields

library :: CombMap Double

neg log prob

expansions :: Double

neg log prob

showLibrary :: Show a => CombMap a -> String

findCLeafInGrammar :: Grammar -> Comb -> Maybe Comb

Is the combinator leaf in the grammar? If so, return it. Else Nothing.

refreshCombFromGrammar :: Grammar -> Comb -> Maybe Comb

Grab the combinator from the grammar. Useful, to get the original type information.

nullGrammar :: Grammar

logsumexp :: (Floating c, Ord c) => [c] -> c

normalizeGrammar :: Grammar -> Grammar

countExpansions :: Comb -> Int

Returns the number of expansions used in this combinator.

countAlts

Arguments

:: [Comb]
a list of combinators (e.g. from a library)

-> Type
a type t

-> CombMap Int
a map of combinators that return type t |

combineGrammars :: (Grammar, Double) -> (Grammar, Double) -> Grammar

Combine two grammars weighted by the number of observations (or pseudo-observations) each has.

bernLogProb :: Int -> Int -> Double

showCombWType :: [Comb] -> String

estimateGrammar

Arguments

:: Grammar
prior

-> Double
number of pseudo-observations by which to weight the prior

-> CombMap [Type]
primitive combinators and their occurance counts

-> [(Task, Comb)]

-> Grammar

calcLogProb :: Grammar -> Type -> Comb -> Double

Returns the log probability of using the given combinator when prompted by the given type, as prescribed by the grammar.

truncateGrammar :: Grammar -> Int -> Grammar