Strappy.Type

Description

This module defines a Hindley-Milner type system. It is based on Mark P. Jone's paper Typing Haskell in Haskell http://web.cecs.pdx.edu/~mpj/thih/.

Types

data Kind

Constructors

Star

Kfun Kind Kind

data Type

Constructors

TVar TyVar

TCon TyCon

TAp Type Type

data TyVar

Constructors

TyVar Id Kind

data TyCon

Constructors

TyCon Id Kind

type Subst

Arguments

= [(TyVar, Type)]
list of type variable - type associations

Substitutions

Functions

(->-) :: Type -> Type -> Type

Function application.

>>> tInt ->- tInt
(Int -> Int)  

freshInst :: Monad m => Type -> StateT Int m Type

Return a type where each type variable is replaced with a new, unbound, type variable.

newTVar :: Monad m => Kind -> StateT Int m Type

mkTVar :: Int -> Type

Convenience function that returns a singleton type consisting of a type variable with specified integer id.

>>> mkTVar 5
v5

mgu :: Type -> Type -> Maybe Subst

apply :: Types t => Subst -> t -> t

merge :: Subst -> Subst -> Maybe Subst

match :: Type -> Type -> Maybe Subst

eqModTyVars

Arguments

:: Type
t1

-> Type
t2

-> Bool

Type equality modulo type variables. Returns true if the mgu susbtitution from t1 to t2 is just a renaming of type variables.

fromType :: Type -> Type

Get the source type of a function type (if it is not a function type then error).

toType :: Type -> Type

Get the target type of a function type (if it is not a function type then error).

tv :: Types t => t -> [TyVar]

isTAp :: Type -> Bool

Returns true if type is a function type.

isTVar :: Type -> Bool

Returns true if type is a singleton type variable.

enumId :: Int -> Id

readId :: String -> Int

nullSubst :: Subst

The empty substitution.

(-->)

Arguments

:: TyVar
u, a type variable

-> Type
t, a type

-> Subst

Returns the substitution of u for t.

Combinator Types

tChar :: Type

Useful types.

tInt :: Type

tBool :: Type

tArrow :: Type

tList :: Type

tMaybe :: Type

tPair :: Type

tTriple :: Type

tQuad :: Type

tQuint :: Type